# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopping List API",
        "description": "Shopping List is a simple project to manage shopping lists via API REST.<br>This is allow shopping list to be created, updated and deleted using request in JSON format and standard HTTP verbs.<br>Bellow, you can check the docs out and learn how to use.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obtain user token",
                "description": "This API helps you to log in and get user token.<br><br>Demo user created<br>Email: contato@matheusmarzochi.com.br<br>Password:password",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "default": "contato@matheusmarzochi.com.br"
                                    },
                                    "password": {
                                        "type": "string",
                                        "default": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obtain user logged information",
                "description": "This API helps you to get more information about user logged in",
                "operationId": "f0b15943f55143d36f606a2dbd1d2366",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log user out",
                "description": "This API helps you to log out",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Token",
                "description": "This API helps you to refresh user token",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List all products",
                "description": "This API helps you to view all the products.",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a product",
                "description": "This API helps you to create a new product.",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shopping_list_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "qty": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve a product",
                "description": "This API lets you retrieve and view a specific product by id.",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier for the resource",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "This API helps you to update a product.",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier for the resource",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shopping_list_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "qty": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "This API helps you delete a product",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier for the resource",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shopping-lists": {
            "get": {
                "tags": [
                    "Shopping Lists"
                ],
                "summary": "List all shopping lists",
                "description": "This API lets you retrieve all shopping lists",
                "operationId": "cffb8e2c8adc0f396fc8baa4ad1b36f5",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shopping Lists"
                ],
                "summary": "Create a shopping list",
                "description": "This API helps you to create a new shopping list",
                "operationId": "c13358e3693a749ca53b5d3b0f17590f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shopping-lists/{id}": {
            "get": {
                "tags": [
                    "Shopping Lists"
                ],
                "summary": "Retrieve a shopping list",
                "description": "This API lets you retrieve a shopping list by id",
                "operationId": "8b3695ebb5273517b6a48853a30a9e4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier for the resource",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Shopping Lists"
                ],
                "summary": "Update a shopping list",
                "description": "This API lets you make changes to a shopping list",
                "operationId": "d942635cf667c5a76533c4d895a5d2ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier for the resource",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Shopping Lists"
                ],
                "summary": "Delete a shopping list",
                "description": "This API helps you delete a shopping list",
                "operationId": "a2a8bfb76202aaa51395627b8c816c38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier for the resource",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shopping-lists/{id}/clone": {
            "post": {
                "tags": [
                    "Shopping Lists"
                ],
                "summary": "Clone  a shopping list",
                "description": "This API lets you to clone a shopping list",
                "operationId": "153bdc89145645e38392d9d18de59a1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier for the resource",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication"
        },
        {
            "name": "Shopping Lists"
        },
        {
            "name": "Products"
        }
    ]
}